refactor index-svc-model
